requirements.txt
Django==1.11
django-cors-headers==2.4.0
djangorestframework==3.8.2
mongoengine==0.15.3
pkg-resources==0.0.0
pymongo==3.7.1
pytz==2018.5
six==1.11.0
